exercices/19-add-contact-form/src/app/add-contact/add-contact.component.spec.ts
import { async, ComponentFixture, TestBed } from '@angular/core/testing';

import { AddContactComponent } from './add-contact.component';
import { MatInputModule } from '@angular/material/input';
import { MatCardModule } from '@angular/material/card';
import { MatButtonModule } from '@angular/material/button';
import { ReactiveFormsModule } from '@angular/forms';
import { NoopAnimationsModule } from '@angular/platform-browser/animations';

describe('AddContactComponent', () => {
  let component: AddContactComponent;
  let fixture: ComponentFixture<AddContactComponent>;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      imports: [
        ReactiveFormsModule,
        MatInputModule,
        MatCardModule,
        MatButtonModule,
        NoopAnimationsModule
      ],
      declarations: [AddContactComponent]
    })
      .compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(AddContactComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should create the form correctly', () => {
    expect(component.addContactForm.get('firstName').value).toBe('');
    expect(component.addContactForm.get('lastName').value).toBe('');
    expect(component.addContactForm.get('phone').value).toBe('');
    expect(component.addContactForm.get('email').value).toBe('');
  });
});
